backend/requirements.txt
Django==3.2.16
django-filter==23.3
djangorestframework==3.14.0
psycopg2-binary==2.9.3
djoser==2.2.0
Pillow==10.0.1
gunicorn==20.1.0
django-colorfield==0.10.1
